implemented sendemail and s3 functionality in UtilController.class
