+++ src/main/java/com/src/constant/AppConfig.java
@@ -3,20 +3,44 @@ package com.src.constant;
 public class AppConfig {
 
 	public static final String CHECK_USERNAME_USERNOTFOUND_ERRORMSG = "checkUsername_usernotfound_errorMsg";
+
 	public static final String SAVEUSER_UNABLE_TO_SAVE_ERRORMSG = "saveUser_unabletosave_errorMsg";
+
 	public static final String FINDBYUSERNAME_USERNOTFOUND_ERRORMSG = "findByUsername_usernotfound_errorMsg";
+
 	public static final String GETUSERBYUSERID_USERNOTFOUND_ERRORMSG = "getUserByUserId_usernotfound_errorMsg";
+
 	public static final String GETALLUSERS_NOUSERSFOUND_ERRORMSG = "getAllUsers_nousersfound_errorMsg";
+
 	public static final String GETALLUSERSBYROLE_NOADMINUSERSFOUND_ERRORMSG = "getAllUsersByRole_noadminusersfound_errorMsg";
+
 	public static final String GETALLUSERSNOTNOTIFICATEDTOTEAM_NOUSERSFOUND_ERRORMSG = "getAllUsersNotNotificatedToTeam_nousersfound_errorMsg";
+
 	public static final String SAVEORUPDATEUSERDETAILS_UNABLETOUPDATE_ERRORMSG = "saveorupdateUserDetails_unabletoupdate_errorMsg";
 
 	public static final String REFERNCELOOKUP_KEY_ERRORMSG = "referencelookup_key_errorMsg";
+
 	public static final String REFERNCELOOKUP_ERRORMSG = "referencelookup_errorMsg";
+
 	public static final String REFERNCELOOKUPMAPPING_REFID_ERRORMSG = "referencelookupmapping_refId_errorMsg";
+
 	public static final String REFERNCELOOKUPMAPPINGSUBCATEGORIES_MAPID_ERRORMSG = "referencelookupmappingcategories_mapId_errorMsg";
 
 	public static final String LOOKUPTEMPALTE_NAME_ERRORMSG = "lookuptemplate_name_errorMsg";
+
+	/*
+	 * Define below are the Email Subject and Short Key Constants
+	 */
+	public final static String EMAIL_SUBJECT_SOMETHINGWENTWRONG = "Something Went Wrong. Urgent Attention Needed";
+
+	public final static String EMAIL_SHORTKEY_SOMETHINGWENTWRONG = "autogen4";
+
+	public final static String EMAIL_SUBJECT_VERIFYEMAILADDRESS = "Verify your email";
+
+	public final static String EMAIL_SHORTKEY_VERIFYEMAILADDRESS = "eventgen15";
+	
+	public final static String EMAIL_SENT_FROMUSER = "team.spprt2019@gmail.com";
+
 
 	private AppConfig() {
 	}
