+++ src/main/java/com/src/restcontroller/UtilityController.java
@@ -6,6 +6,7 @@ import org.json.JSONException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -14,17 +15,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 import com.src.entity.UtilEntity;
+import com.src.helper.SendEmailHelper;
 
 @Controller
-public class UtilController extends AbstractRestManager {
-	
-	final Logger logger = LoggerFactory.getLogger(UtilController.class);
+public class UtilityController extends AbstractRestManager {
 
-	@RequestMapping(value = "/sendEmail/", method = RequestMethod.POST)
-	public ResponseEntity<UtilEntity> sendEmail(@RequestBody UtilEntity utilEntity) throws JSONException {
-		UtilEntity utilEntityResponse = utilService.sendEmail(utilEntity);
-		return new ResponseEntity<UtilEntity>(utilEntityResponse, HttpStatus.OK);
-	}
+	final Logger logger = LoggerFactory.getLogger(UtilityController.class);
 
 	@RequestMapping(value = "/uploadAvatarsInS3/{inputFile}/{userid}", method = RequestMethod.GET)
 	public ResponseEntity<String> uploadAvatarsInS3(@PathVariable("inputFile") File inputFile,
@@ -46,6 +42,22 @@ public class UtilController extends AbstractRestManager {
 		String widgetURL = utilService.uploadWidgetPicsInS3(inputFile, userid);
 		return new ResponseEntity<String>(widgetURL, HttpStatus.OK);
 
+	}
+
+	@RequestMapping(value = "/autoSendEmail/", method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE)
+	public ResponseEntity<UtilEntity> autoSendEmail(@RequestBody UtilEntity utilEntity) throws JSONException {
+		UtilEntity utilEntityResponse = utilService.sendEmail(utilEntity);
+		return new ResponseEntity<UtilEntity>(utilEntityResponse, HttpStatus.OK);
+	}
+
+	@RequestMapping(value = "/sendEmail/", method = RequestMethod.PUT)
+	public ResponseEntity<UtilEntity> sendEmail(@RequestBody Object obj, @RequestBody String templateURL,
+			@RequestBody String shortKey) throws JSONException {
+
+		SendEmailHelper emailHelper = new SendEmailHelper();
+		UtilEntity utilEntity = emailHelper.returnUtilEntityObjByShortKey(obj, templateURL, shortKey);
+		UtilEntity utilEntityResponse = utilService.sendEmail(utilEntity);
+		return new ResponseEntity<UtilEntity>(utilEntityResponse, HttpStatus.OK);
 	}
 
 }
